requirements.txt
selenium==4.15.2
beautifulsoup4==4.12.2
requests==2.31.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
pandas==2.1.3
openpyxl==3.11.0
playwright==1.40.0
lxml==4.9.3
aiohttp==3.9.1
asyncio-contextmanager==1.0.0
email-validator==2.1.0
